# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Rest Api for Fluffy Doodle",
        "title": "Fluffy Doodle REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/user/avatar": {
            "post": {
                "description": "Upload user avatar to S3",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Upload avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Avatar file",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/parent-sign-up": {
            "post": {
                "description": "Parent creates account and links to student using student code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Parent registration by student code",
                "parameters": [
                    {
                        "description": "Parent registration data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.ParentSignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/parent/child-progress": {
            "get": {
                "description": "Returns current child progress for authenticated parent",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Parent child progress",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ParentChildProgressResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "description": "Получение профиля текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/student-invitation": {
            "post": {
                "description": "Teacher creates student account and receives student code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Invite student",
                "parameters": [
                    {
                        "description": "Student registration data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/presenter.StudentInviteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.AuthSwaggerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "presenter.AuthSwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
        "presenter.AuthSwaggerSuccessResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "presenter.ParentChildProgressResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "integer",
                    "example": 4
                },
                "stars": {
                    "type": "integer",
                    "example": 18
                },
                "studentCode": {
                    "type": "string",
                    "example": "STU-8A7KQ21M"
                },
                "studentId": {
                    "type": "integer",
                    "example": 12
                },
                "studentLogin": {
                    "type": "string",
                    "example": "petya_ivanov"
                },
                "studentName": {
                    "type": "string",
                    "example": "Петя Иванов"
                }
            }
        },
        "presenter.ParentSignUpRequest": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string",
                    "example": "Мария Иванова"
                },
                "login": {
                    "type": "string",
                    "example": "parent01"
                },
                "password": {
                    "type": "string",
                    "example": "superSecret123"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "+79001234567"
                },
                "studentCode": {
                    "type": "string",
                    "example": "STU-8A7KQ21M"
                }
            }
        },
...
        "presenter.SignUpRequest": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "presenter.StudentInviteResponse": {
            "type": "object",
            "properties": {
                "studentCode": {
                    "type": "string",
                    "example": "STU-8A7KQ21M"
                }
            }
        },
...
        "presenter.UserProfileResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://minio.local/avatar_1.png"
                },
                "fullName": {
                    "type": "string",
                    "example": "Иван Иванов"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "login": {
                    "type": "string",
                    "example": "teacher01"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "+79001234567"
                },
                "roleId": {
                    "type": "integer",
                    "example": 2
                }
            }
        }
    }
}